+++ singletons/interface.py
@@ -1,12 +1,18 @@
 # -*- coding: utf-8 -*-
 
 import os
+import sys
 import glob
 import xml.etree.ElementTree as ElementTree
 from collections import namedtuple
-from typing import List
+from typing import List, Optional
 
 from singletons.config import config
+
+if getattr(sys, 'frozen', False):
+    base_path = sys._MEIPASS # type: ignore
+else:
+    base_path = os.path.abspath(".")
 
 
 class Lang(namedtuple('Lang', 'code name items authors version')):
@@ -22,11 +28,12 @@ class Interface:
 
     def __init__(self) -> None:
         self.__languages = {}
-        self.__current = None
+        self.__current: Optional[Lang] = None
         self.__load()
 
     def __load(self) -> None:
-        files = glob.glob(os.path.join('./prefs/interface', '*.xml'))
+        interface_folder_path = os.path.join(base_path, 'prefs', 'interface')
+        files = glob.glob(os.path.join(interface_folder_path, '*.xml'))
 
         for f in files:
             with open(f, 'r', encoding='utf-8') as fp:
@@ -66,17 +73,18 @@ class Interface:
         self.reload()
 
     def reload(self) -> None:
-        self.__current = self.__languages.get(config.value('interface', 'language'), None)
+        lang_code = str(config.value('interface', 'language'))
+        self.__current = self.__languages.get(lang_code, None)
 
     def text(self, k: str, v: str) -> str:
         return self.__current.get(k, v) if isinstance(self.__current, Lang) else v
 
     @property
-    def authors(self) -> str:
+    def authors(self) -> Optional[str]:
         return self.__current.authors if isinstance(self.__current, Lang) else None
 
     @property
-    def version(self) -> str:
+    def version(self) -> Optional[str]:
         return self.__current.version if isinstance(self.__current, Lang) else None
 
     @property
@@ -86,7 +94,7 @@ class Interface:
     @property
     def current_index(self) -> int:
         keys = list(self.__languages.keys())
-        interface_value = config.value('interface', 'language')
+        interface_value = str(config.value('interface', 'language'))
         if interface_value in keys:
             return keys.index(interface_value)
         return 0