+++ singletons/languages.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 import os
+import sys
 from collections import namedtuple
 import xml.etree.ElementTree as ElementTree
 
@@ -8,6 +9,11 @@ from singletons.config import config
 
 
 Language = namedtuple('Language', 'locale code google deepl')
+
+if getattr(sys, 'frozen', False):
+    base_path = sys._MEIPASS # type: ignore
+else:
+    base_path = os.path.abspath(".")
 
 
 class Languages:
@@ -18,7 +24,7 @@ class Languages:
         self.__load()
 
     def __load(self):
-        languages_file = os.path.abspath('./prefs/languages.xml')
+        languages_file = os.path.join(base_path, 'prefs', 'languages.xml')
 
         if not os.path.exists(languages_file):
             return
@@ -45,17 +51,19 @@ class Languages:
         return list(self.__locales.keys())
     
     @property
-    def source(self) -> Language:
-        return self.by_locale(config.value('translation', 'source'))
+    def source(self) -> 'Language | None': 
+        locale = str(config.value('translation', 'source'))
+        return self.by_locale(locale)
 
     @property
-    def destination(self) -> Language:
-        return self.by_locale(config.value('translation', 'destination'))
+    def destination(self) -> 'Language | None':
+        locale = str(config.value('translation', 'destination'))
+        return self.by_locale(locale)
 
-    def by_locale(self, locale: str) -> Language:
+    def by_locale(self, locale: str) -> 'Language | None':
         return self.__locales.get(locale)
 
-    def by_code(self, code: str) -> Language:
+    def by_code(self, code: str) -> 'Language | None':
         return self.__codes.get(code)