+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Version 1.3.2 - 2025-06-02
+
+- Adding Indonesian language support
+- Fix “FileNotFoundError” in the prefs directory
+
 ## Version 1.3.1 - 2024-08-02
 
 - Added German language